requirements.api.txt
fastapi
uvicorn[standard]
requests
mysql-connector-python
pika
openai
python-dotenv
